AR Foundation and building settings optimized
Build settings have now been set up to handle mobile. AR foundation is now installed to the project ready for Android (This is because I do not own an iOS product). However the architecture can handle platform changes to account for that if I need to build for iOS devices.